水道料金ページのリンクを新しいものに変更
