+++ pages/charge/index.vue
@@ -1,783 +1,1459 @@
 <template>
-  <div>
-    <BreadCrumb :breadcrumbs="breadcrumbs" />
-    <v-container fluid class="charge-intro">
-      <section>
-        <v-row justify="center">
-          <v-col cols="12" md="8">
-            <v-row class="my-2">
-              <div
-                class="text-h4 text-md-h2 mx-2 my-3 font-weight-bold primary--text"
+  <div class="p-3 md:p-4 max-w-6xl mx-auto bg-gray-50 rounded-lg shadow">
+    <h1 class="text-2xl font-bold text-center mb-6 text-main">
+      水道料金分析ページ
+    </h1>
+
+    <!-- タブメニュー -->
+    <div class="mb-6">
+      <!-- PC用タブメニュー -->
+      <div
+        v-if="!isMobile"
+        class="radio-tabs bg-white p-2 rounded-lg shadow-sm grid grid-cols-3 gap-2"
+      >
+        <label
+          v-for="tab in ['compare', 'graph', 'ranking', 'averageRanking']"
+          :key="tab"
+          class="radio-tab cursor-pointer p-3 rounded-md text-center font-medium flex items-center justify-center transition-all"
+          :class="
+            activeTab === tab
+              ? 'bg-main text-white shadow-sm'
+              : 'bg-gray-100 hover:bg-gray-200'
+          "
+        >
+          <input v-model="activeTab" type="radio" :value="tab" class="hidden" />
+          <span>
+            {{
+              tab === 'compare'
+                ? '水道料金比較シミュレーター'
+                : tab === 'graph'
+                ? '都道府県内比較グラフ'
+                : tab === 'ranking'
+                ? '全国自治体水道料金ランキング'
+                : '都道府県平均水道料金ランキング'
+            }}
+          </span>
+        </label>
+      </div>
+
+      <!-- モバイル用プルダウンメニュー -->
+      <div v-if="isMobile" class="px-2">
+        <select
+          v-model="activeTab"
+          class="w-full p-3 bg-white border border-gray-300 rounded-lg shadow-sm font-medium text-gray-700 focus:outline-none focus:ring-2 focus:ring-main focus:border-main"
+        >
+          <option value="compare">水道料金比較シミュレーター</option>
+          <option value="graph">都道府県内比較グラフ</option>
+          <option value="ranking">全国自治体水道料金ランキング</option>
+          <option value="averageRanking">都道府県平均水道料金ランキング</option>
+        </select>
+      </div>
+    </div>
+
+    <!-- 水道料金比較シミュレーター -->
+    <div
+      v-if="activeTab === 'compare'"
+      class="bg-white p-4 md:p-6 rounded-lg shadow-sm"
+    >
+      <h2 class="text-xl font-semibold mb-4 text-main">
+        水道料金比較シミュレーター
+      </h2>
+
+      <!-- 自治体選択部分を横並びに -->
+      <div class="comparison-container mb-6">
+        <!-- 左側の自治体 -->
+        <div class="w-full bg-light-blue p-4 rounded-lg">
+          <div class="space-y-4">
+            <div>
+              <label class="block mb-2 font-medium">都道府県を選択（1）</label>
+              <select
+                v-model="selectedPrefecture"
+                class="w-full p-2 border rounded"
               >
-                水道料金について
-              </div>
-            </v-row>
-            <v-divider></v-divider>
-            <v-row class="my-2">
-              <v-col cols="12" md="5">
-                <h1>はじめに</h1>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;水道料金が自治体によって異なることは、多くの方がご存知の方かと思います。
-                実際にはどれだけ違いがあるのでしょうか？<br /><br />
+                <option
+                  v-for="pref in availablePrefectures"
+                  :key="pref"
+                  :value="pref"
+                >
+                  {{ pref }}
+                </option>
+              </select>
+            </div>
+            <div>
+              <label class="block mb-2 font-medium">自治体を選択（1）</label>
+              <select v-model="selectedCity" class="w-full p-2 border rounded">
+                <option
+                  v-for="city in availableCitiesInPrefecture"
+                  :key="city"
+                  :value="city"
+                >
+                  {{ city }}
+                </option>
+              </select>
+            </div>
+            <div class="bg-white p-4 rounded-lg">
+              <h3 class="text-lg font-medium mb-2 text-main">計算結果（1）</h3>
+              <p class="text-3xl font-bold text-main-dark">
+                {{ bill.toLocaleString() }} 円
+              </p>
+            </div>
+          </div>
+        </div>
 
-                &emsp;自治体ごとの水道料金をランキングしたWebサイトはありますが、詳細なデータを確認することはできません。
-                このWebサイトでは、全国の自治体の水道料金や都道府県の平均水道料金をランキング形式で表示しています。<br /><br />
-
-                &emsp;一般的な水道使用量だけではなく、口径や使用水量による料金の差、基本料金と超過料金のデータについても掲載していますので、
-                ご自身の住んでいるまちや興味のあるまちの情報をチェックしてみてください。<br />
-              </div>
-            </v-row>
-            <v-row class="mt-2">
-              <v-col cols="12" md="5">
-                <h1>目次</h1>
-              </v-col>
-            </v-row>
-            <v-row class="mb-5">
-              <v-col cols="12" md="10" class="ml-10">
-                <ul color="rgba(242 242 242)">
-                  <li>
-                    <nuxt-link v-scroll-to="'#intro'" to>
-                      水道事業と水道料金
-                    </nuxt-link>
-                  </li>
-                  <li>
-                    <nuxt-link v-scroll-to="'#how'" to>
-                      水道料金算定の仕組み
-                    </nuxt-link>
-                    <ul>
-                      <li>総括原価方式</li>
-                      <li>水道料金体系</li>
-                    </ul>
-                  </li>
-                  <li>
-                    <nuxt-link v-scroll-to="'#variant'" to>
-                      口径と使用量
-                    </nuxt-link>
-                    <ul>
-                      <li>口径について</li>
-                      <li>使用水量について</li>
-                      <li>（例）口径と使用水量の違いによる料金の違い</li>
-                    </ul>
-                  </li>
-                  <li>
-                    <nuxt-link v-scroll-to="'#data'" to>
-                      水道料金データの見方
-                    </nuxt-link>
-                  </li>
-                </ul>
-              </v-col>
-            </v-row>
-            <v-divider></v-divider>
-            <v-row id="intro" class="my-2">
-              <v-col cols="12" md="5">
-                <h1>水道事業と水道料金</h1>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;水道料金を決定している水道事業とはどんな事業なのでしょうか？
-                水道事業は、地方公共団体が住民の福祉の増進を目的として設定し、
-                経営する地方公営企業として運営されています。<br />
-              </div>
-              <div class="box1">
-                <p>
-                  &emsp;地方公共団体は、一般的な行政活動の他、水の供給や公共輸送の確保、医療の提供、
-                  下水の処理など地域住民の生活や地域の発展に不可欠なサービスを提供する様々な事業活動を行っている。
-                  こうした事業を行うために地域公共団体が経営する企業活動を総称して「地方公営企業」と呼んでおり、
-                  水道事業、下水道事業、交通事業、病院事業等がその代表的なものであるが、
-                  その他にも、電気（卸売）・ガス事業や土地造成事業を行うなど、その事業種別は多種多様である。
-                </p>
-              </div>
-              <div>
-                参考：
-                <a
-                  href="https://www.soumu.go.jp/main_sosiki/c-zaisei/kouei.html"
-                  target="_blank"
-                >
-                  総務省「地方公営企業等」
-                </a>
-              </div>
-              <div class="my-10">
-                &emsp;地方公営企業である水道事業は、法律で「<strong>地方公営企業の経営に伴う収入をもつて充てなければならない</strong>」（
-                <a
-                  href="https://elaws.e-gov.go.jp/document?lawid=327AC0000000292#Mp-At_17_2"
-                  target="_blank"
-                  >地方公営企業法第17条の2第2項</a
-                >
-                ）と規定されており、基本的には水道料金等の事業収入で経営しなければならない原則があります。
-              </div>
-              <div>
-                &emsp;また、水道料金は、「<strong
-                  >公正妥当なものでなければならず、かつ、能率的な経営の下における適正な原価を基礎とし、
-                  地方公営企業の健全な運営を確保することができるものでなければならない</strong
-                >（
-                <a
-                  href="https://elaws.e-gov.go.jp/document?lawid=327AC0000000292#Mp-At_21"
-                  target="_blank"
-                  >地方公営企業法第21条第2項</a
-                >
-                ）」、「<strong>定率または定額をもって明確に定められていること</strong>」、
-                「<strong
-                  >特定の者に対して不当な差別的取り扱いをするものでないこと </strong
-                >」、等（
-                <a
-                  href="https://elaws.e-gov.go.jp/document?lawid=332AC0000000177#Mp-At_14"
-                  target="_blank"
-                  >水道法第14条第2項各号</a
-                >
-                ）、公正妥当であることが求められています。
-              </div>
-              <div class="my-10">
-                &emsp;では、水道料金はどのように決定されているのでしょうか？
-              </div>
-            </v-row>
-
-            <v-divider></v-divider>
-            <v-row id="how" class="my-2">
-              <v-col cols="12" md="5">
-                <h1>水道料金算定の仕組み</h1>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                参考：日本水道協会「
-                <a
-                  href="http://www.jwwa.or.jp/houkokusyo/pdf/suidou_santei/suidou_santei_02.pdf"
-                  target="_blank"
-                >
-                  水道料金算定要領
-                </a>
-                、水道料金改定業務の手引き」
-              </div>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <h2 class="primary--text">総括原価</h2>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div class="my-1">
-                &emsp;水道事業では「総括原価方式」という方法で水道料金が決定されています。
-                水道事業における総括原価方式とは、発生した営業費用（人件費、修繕費、減価償却費等）
-                に資本費用（支払利息、資産維持費）を加え、水道料金収入以外の収入を差し引いて料金を算定する方法です。<br /><br />
-
-                &emsp;資産維持費とは施設の改修・更新に必要となる費用とされています。
-                これは、「事業運営にかかるコスト」と「かかるであろうコスト」を水道料金収入で賄うように
-                料金水準が設定されていることを意味します。
-              </div>
-            </v-row>
-            <v-row class="my-3 mx-3">
-              <v-img :src="full_cost" max-width="100%"></v-img>
-            </v-row>
-            <br />
-            <v-row class="my-2 mx-3">
-              <h2 class="primary--text">水道料金体系</h2>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div class="my-2">
-                &emsp;総括原価方式で必要な料金収入の総額が決定された後、水道料金体系が設定されます。
-                水道料金体系は、使用している水道管の太さ（口径）、使用水量、用途等によって各自治体で定められています。
-                水道料金の体系は、大きく<strong>「用途別」</strong>と<strong>「口径別」</strong>に分けられ、
-                一般的に固定的にかかる費用に当たる<strong>「基本料金」</strong>と
-                使用した水量に応じて負担する費用にあたる<strong>「従量料金」</strong>で構成される
-                <strong>「二部料金性」</strong>が採用されています。
-              </div>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <v-img :src="charge_system" max-width="50%"></v-img>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <h3>全国の事業体における水道料金体系別割合 <br /></h3>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;水道料金体系は普段意識することは少ないかもしれませんが、下の表でまとめたように、
-                区分が「用途別」なのか「口径別」なのか「その他」なのか、
-                基本料金において「基本水量あり」なのか「基本水量なし」なのか等、
-                水道事業体によって様々な体系が存在します。
-              </div>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                出典：日本水道協会「水道料金表（令和3年4月1日現在）」<br />
-              </div>
-            </v-row>
-            <v-row class="my-2">
-              <v-col cols="12">
-                <table
-                  border="1"
-                  style="border-collapse: collapse"
-                  bgcolor="#FFFFFF"
-                  width="90%"
-                >
-                  <tr>
-                    <th valign="bottom">区分</th>
-                    <th valign="bottom">基本料金（準備料金）</th>
-                    <th valign="bottom">事業体数</th>
-                    <th valign="bottom">比率（%）</th>
-                    <th valign="bottom">小計</th>
-                  </tr>
-                  <tr bgcolor="#BDD7EE">
-                    <th rowspan="3" align="center">用途別</th>
-                    <td valign="bottom">用途別基本料金<br />基本水量あり</td>
-                    <td align="right" valign="bottom">367</td>
-                    <td align="right" valign="bottom">29.1</td>
-                    <td rowspan="3" align="right" valign="bottom">380</td>
-                  </tr>
-                  <tr bgcolor="#BDD7EE">
-                    <td valign="bottom">単一基本料金<br />基本水量あり</td>
-                    <td align="right" valign="bottom">12</td>
-                    <td align="right" valign="bottom">1.0</td>
-                  </tr>
-                  <tr bgcolor="#BDD7EE">
-                    <td valign="bottom">単一基本料金<br />基本水量なし</td>
-                    <td align="right" valign="bottom">1</td>
-                    <td align="right" valign="bottom">0.1</td>
-                  </tr>
-                  <tr bgcolor="#F8CBAD">
-                    <th rowspan="2" align="center">口径別</th>
-                    <td valign="bottom">基本水量無し<br />基本料金</td>
-                    <td align="right" valign="bottom">323</td>
-                    <td align="right" valign="bottom">25.6</td>
-                    <td rowspan="2" align="right" valign="bottom">752</td>
-                  </tr>
-                  <tr bgcolor="#F8CBAD">
-                    <td valign="bottom">基本水量付<br />基本料金</td>
-                    <td align="right" valign="bottom">429</td>
-                    <td align="right" valign="bottom">34.0</td>
-                  </tr>
-                  <tr bgcolor="#DBDBDB">
-                    <th rowspan="3" align="center">その他</th>
-                    <td valign="bottom">基本水量付<br />単一基本料金</td>
-                    <td align="right" valign="bottom">116</td>
-                    <td align="right" valign="bottom">9.2</td>
-                    <td rowspan="3" align="right" valign="bottom">129</td>
-                  </tr>
-                  <tr bgcolor="#DBDBDB">
-                    <td valign="bottom">基本水量無し<br />単一基本料金</td>
-                    <td align="right" valign="bottom">13</td>
-                    <td align="right" valign="bottom">1.0</td>
-                  </tr>
-                  <tr bgcolor="#DBDBDB">
-                    <td valign="bottom">基本水量無し<br />基本料金なし</td>
-                    <td align="right" valign="bottom">0</td>
-                    <td align="right" valign="bottom">0.0</td>
-                  </tr>
-                  <tr>
-                    <td colspan="2" valign="bottom">合計</td>
-                    <td align="right" valign="bottom">1261</td>
-                    <td align="right" valign="bottom">100</td>
-                    <td align="right" valign="bottom">1261</td>
-                  </tr>
-                </table>
-              </v-col>
-              <v-col cols="12">
-                <v-img :src="ratio_system" max-width="90%"></v-img>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3 mt-10">
-              <h3>全国の料金体系別の推移 <br /></h3>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;下の表とグラフは、水道料金体系の区分別の事業対数及びその割合の推移を示しています。
-                昭和年代は「用途別」が多かったようですが、「口径別」が増加し、令和3年度には約60%になっています。
-              </div>
-            </v-row>
-            <v-row class="my-2 mx-3 table-wrap">
-              <table
-                border="1"
-                style="border-collapse: collapse"
-                bgcolor="#FFFFFF"
-                width="90%"
+        <!-- 右側の自治体 -->
+        <div class="w-full bg-light-orange p-4 rounded-lg">
+          <div class="space-y-4">
+            <div>
+              <label class="block mb-2 font-medium">都道府県を選択（2）</label>
+              <select
+                v-model="comparePrefecture"
+                class="w-full p-2 border rounded"
               >
-                <tr>
-                  <th colspan="2" rowspan="2" align="center">区分</th>
-                  <th rowspan="2" width="60" align="center">S40</th>
-                  <th rowspan="2" width="60" align="center">S60</th>
-                  <th rowspan="2" width="60" align="center">H17</th>
-                  <th rowspan="2" width="60" align="center">H21</th>
-                  <th rowspan="2" width="60" align="center">H25</th>
-                  <th rowspan="2" width="60" align="center">H29</th>
-                  <th rowspan="2" width="60" align="center">R3</th>
-                </tr>
-                <tr></tr>
-                <tr bgcolor="#BDD7EE">
-                  <th rowspan="2" width="70" align="center">用途別</th>
-                  <td width="80" align="center">事業体数</td>
-                  <td align="right">1,095</td>
-                  <td align="right">868</td>
-                  <td align="right">613</td>
-                  <td align="right">471</td>
-                  <td align="right">431</td>
-                  <td align="right">404</td>
-                  <td align="right">380</td>
-                </tr>
-                <tr bgcolor="#BDD7EE">
-                  <td align="center">割合</td>
-                  <td align="right">99.0%</td>
-                  <td align="right">47.1%</td>
-                  <td align="right">38.7%</td>
-                  <td align="right">35.8%</td>
-                  <td align="right">33.7%</td>
-                  <td align="right">31.8%</td>
-                  <td align="right">30.1%</td>
-                </tr>
-                <tr bgcolor="#F8CBAD">
-                  <th rowspan="2" align="center">口径別</th>
-                  <td align="center">事業体数</td>
-                  <td align="right">11</td>
-                  <td align="right">705</td>
-                  <td align="right">783</td>
-                  <td align="right">697</td>
-                  <td align="right">708</td>
-                  <td align="right">725</td>
-                  <td align="right">752</td>
-                </tr>
-                <tr bgcolor="#F8CBAD">
-                  <td align="center">割合</td>
-                  <td align="right">1.0%</td>
-                  <td align="right">38.3%</td>
-                  <td align="right">49.4%</td>
-                  <td align="right">53.0%</td>
-                  <td align="right">55.4%</td>
-                  <td align="right">57.1%</td>
-                  <td align="right">59.6%</td>
-                </tr>
-                <tr bgcolor="#DBDBDB">
-                  <th rowspan="2" align="center">その他</th>
-                  <td align="center">事業体数</td>
-                  <td align="right">0</td>
-                  <td align="right">270</td>
-                  <td align="right">190</td>
-                  <td align="right">148</td>
-                  <td align="right">140</td>
-                  <td align="right">140</td>
-                  <td align="right">129</td>
-                </tr>
-                <tr bgcolor="#DBDBDB">
-                  <td align="center">割合</td>
-                  <td align="right">0.0%</td>
-                  <td align="right">14.7%</td>
-                  <td align="right">12.0%</td>
-                  <td align="right">11.2%</td>
-                  <td align="right">10.9%</td>
-                  <td align="right">11.0%</td>
-                  <td align="right">10.2%</td>
-                </tr>
-                <tr>
-                  <th rowspan="2" align="center">合計</th>
-                  <td align="center">事業体数</td>
-                  <td align="right">1,106</td>
-                  <td align="right">1,843</td>
-                  <td align="right">1,586</td>
-                  <td align="right">1,316</td>
-                  <td align="right">1,279</td>
-                  <td align="right">1,269</td>
-                  <td align="right">1,261</td>
-                </tr>
-                <tr>
-                  <td align="center">割合</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                  <td align="right">100.0%</td>
-                </tr>
-              </table>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <v-img :src="change_graph" max-width="90%"></v-img>
-            </v-row>
+                <option
+                  v-for="pref in availablePrefectures"
+                  :key="pref"
+                  :value="pref"
+                >
+                  {{ pref }}
+                </option>
+              </select>
+            </div>
+            <div>
+              <label class="block mb-2 font-medium">自治体を選択（2）</label>
+              <select v-model="compareCity" class="w-full p-2 border rounded">
+                <option
+                  v-for="city in availableCitiesInComparePrefecture"
+                  :key="city"
+                  :value="city"
+                >
+                  {{ city }}
+                </option>
+              </select>
+            </div>
+            <div class="bg-white p-4 rounded-lg">
+              <h3 class="text-lg font-medium mb-2 text-orange-700">
+                計算結果（2）
+              </h3>
+              <p class="text-3xl font-bold text-orange-800">
+                {{ compareBill.toLocaleString() }} 円
+              </p>
+            </div>
+          </div>
+        </div>
+      </div>
 
-            <v-divider></v-divider>
-            <v-row id="variant" class="my-2">
-              <v-col cols="12" md="5">
-                <h1>口径と使用水量</h1>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;水道料金のデータを理解するためには、口径と使用水量の違いを把握することが大切です。
-                ここでは水道料金の違いに影響を与える「口径」と「使用水量」の説明をします。
-              </div>
-            </v-row>
-            <v-row class="mb-2 mt-7 mx-2">
-              <h2 class="primary--text">口径について</h2>
-            </v-row>
-            <v-row class="mb-1 mt-2 mx-3">
-              <h3>口径とは？ <br /></h3>
-            </v-row>
-            <v-row class="mt-1 mb-2 mx-4">
-              <div>
-                &emsp;口径とは水道メーターに接続されている水道管の直径を指します。
-                この口径は、近くまできている水道管の口径や家庭内の蛇口の数によって決まります。
-                家の水道メーターについてる口径を近所まできている水道管よりも大きい口径にすることはできません。
-                一般家庭における口径は、13mm、20mmが用いられています。
-              </div>
-            </v-row>
-            <v-row class="mb-1 mt-4 mx-3">
-              <h3>口径が変わると？ <br /></h3>
-            </v-row>
-            <v-row class="mt-1 mb-5 mx-4">
-              <div>
-                &emsp;口径の大きさによって維持管理にかかる費用が異なるため、
-                多くの事業体では口径によって料金が異なります。
-                一般的には口径が大きくなればなるほど水道料金が高くなる傾向にあります。
-              </div>
-            </v-row>
+      <!-- 使用量設定 -->
+      <div class="bg-gray-50 p-4 rounded-lg mb-6">
+        <label class="block mb-2 font-medium">使用量 (m³)</label>
+        <input
+          v-model="waterUsage"
+          type="range"
+          min="1"
+          max="50"
+          class="w-full mb-2"
+        />
+        <div class="flex justify-between items-center">
+          <span class="text-lg font-medium">{{ waterUsage }} m³</span>
+          <div class="w-32"></div>
+          <button
+            class="px-3 py-2 bg-main text-white rounded text-sm ml-auto"
+            @click="waterUsage = 20"
+          >
+            標準使用量に設定
+          </button>
+        </div>
+      </div>
 
-            <v-row class="mb-2 mt-7 mx-2">
-              <h2 class="primary--text">使用水量について</h2>
-            </v-row>
-            <v-row class="mt-1 mb-2 mx-4">
-              <div>
-                &emsp;本サイトにおける使用水量とは、1ヶ月間に使った水の量を示しています。
-                使用水量に応じた料金は、（使用水量）×（単価）で計算されます。
-                （単価）は事業者によって異なり、使用水量に関わらず一定の単価や、
-                使用水量が多くなるほど単価が高くなる逓増性を採用している事業体があります。<br /><br />
+      <!-- 比較結果 -->
+      <div class="bg-gray-50 p-6 rounded-lg border-2 border-main-light">
+        <h3 class="text-xl font-bold mb-4 text-gray-700 border-b pb-2">
+          料金差の比較結果
+        </h3>
 
-                &emsp;皆さんは毎月どれくらい家庭で水を使用しているか把握していますか？
-                <a
-                  href="https://www.waterworks.metro.tokyo.lg.jp/kurashi/shiyou/jouzu.html"
-                  target="_blank"
-                  >東京都水道局の調査(令和2年)
-                </a>
-                では、世帯人数による平均の使用水量は以下のような結果でした。
-              </div>
-            </v-row>
-            <v-col cols="8" md="4">
-              <v-row class="my-2 mx-3 box1">
-                <ul>
-                  <li>1人世帯：8.1㎥</li>
-                  <li>2人世帯：14.9㎥</li>
-                  <li>3人世帯：19.9㎥</li>
-                  <li>4人世帯：23.1㎥</li>
-                  <li>5人世帯：27.8㎥</li>
-                  <li>6人以上世帯：34.1㎥</li>
-                </ul>
-              </v-row>
-            </v-col>
-            <v-row class="mt-1 mb-1 mx-4">
-              <div>
-                &emsp;皆さんの毎月の使用量は平均と比較してどうでしょうか？
-                本サイトで水道料金を確認する際は平均の使用量や実際の自分の使用量などを参考にご覧ください。<br /><br />
-              </div>
-            </v-row>
-            <v-row class="mt-7 mb-3 mx-2">
-              <h2 class="primary--text">
-                口径と使用水量の違いによる料金の違いの例
-              </h2>
-            </v-row>
-            <v-row class="my-2 mx-2">
-              &emsp;口径と使用水量の違いによる料金の違いを具体的な事例で見てみましょう。
-              まずは使用水量が同じでも口径が違う場合についてみてみます。
-            </v-row>
-            <v-row class="my-2 mx-2">
-              <table
-                border="1"
-                style="border-collapse: collapse"
-                bgcolor="#FFFFFF"
-                cellpadding="100"
+        <div class="text-center">
+          <p class="text-lg text-gray-700 bg-white p-4 rounded-lg shadow-sm">
+            <span class="font-bold text-main">
+              {{ bill > compareBill ? selectedCity : compareCity }}
+            </span>
+            の方が
+            <span class="font-bold text-red-600">
+              {{ Math.abs(bill - compareBill).toLocaleString() }}円
+            </span>
+            高くなっています
+          </p>
+        </div>
+
+        <div class="mt-4 text-center text-gray-600">
+          <p class="text-sm">
+            差額は
+            <span class="font-bold">
+              {{
+                Math.round(
+                  (Math.abs(bill - compareBill) / Math.min(bill, compareBill)) *
+                    100
+                )
+              }}%
+            </span>
+            の違いがあります
+          </p>
+        </div>
+      </div>
+    </div>
+
+    <!-- 都道府県内比較グラフ -->
+    <div
+      v-if="activeTab === 'graph'"
+      class="bg-white p-4 md:p-6 rounded-lg shadow-sm"
+      style="height: auto"
+    >
+      <h2 class="text-xl font-semibold mb-4 text-main">都道府県内比較グラフ</h2>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">都道府県を選択</label>
+        <select
+          v-model="selectedGraphPrefecture"
+          class="w-full p-2 border rounded"
+        >
+          <option
+            v-for="pref in availablePrefectures"
+            :key="pref"
+            :value="pref"
+          >
+            {{ pref }}
+          </option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">使用量を選択</label>
+        <select v-model="selectedUsageFee" class="w-full p-2 border rounded">
+          <option value="10m3">10m³</option>
+          <option value="15m3">15m³</option>
+          <option value="20m3">20m³</option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">自治体名で検索</label>
+        <input
+          v-model="graphSearchCity"
+          type="text"
+          class="w-full p-2 border rounded"
+          placeholder="自治体名を入力してください"
+        />
+        <div class="flex mt-2 space-x-4">
+          <button
+            class="px-6 py-2 bg-main text-white rounded hover:bg-main-dark"
+            @click="searchGraphCity"
+          >
+            検索
+          </button>
+          <button
+            v-if="graphHighlightedCity"
+            class="px-6 py-2 text-white rounded hover:bg-gray-600"
+            style="background-color: #6b7280"
+            @click="clearGraphSearch"
+          >
+            クリア
+          </button>
+        </div>
+      </div>
+
+      <!-- 検索結果の表示 -->
+      <div
+        v-if="graphHighlightedCity && graphHighlightedCityFee"
+        class="mb-3 p-3 bg-blue-50 rounded-lg border border-blue-200"
+      >
+        <p class="text-center font-medium">
+          「{{ graphHighlightedCity }}」の水道料金:
+          <span class="text-xl font-bold text-main"
+            >{{ graphHighlightedCityFee.toLocaleString() }}円</span
+          >
+          ({{ selectedUsageFee }}使用時)
+        </p>
+      </div>
+
+      <!-- モバイル表示の場合は横棒グラフ、PC表示の場合は縦棒グラフ -->
+      <div v-if="!isMobile" style="min-height: 500px">
+        <GChart
+          type="ColumnChart"
+          :data="chartData"
+          :options="{
+            title: '都道府県別水道料金',
+            chartArea: { width: '70%', height: '70%' },
+            height: 600,
+            hAxis: {
+              title: '自治体',
+              slantedText: true,
+              slantedTextAngle: 45,
+              textStyle: {
+                fontSize: 12,
+              },
+              showTextEvery: 1,
+            },
+            vAxis: {
+              title: '料金 (円)',
+              minValue: 0,
+            },
+            bar: { groupWidth: '60%' },
+            legend: { position: 'none' },
+          }"
+        />
+      </div>
+      <div v-if="isMobile" style="min-height: 500px">
+        <GChart
+          type="BarChart"
+          :data="chartData"
+          :options="{
+            title: '都道府県別水道料金',
+            chartArea: { width: '65%', height: '80%' },
+            height: 500,
+            hAxis: {
+              title: '料金 (円)',
+              minValue: 0,
+            },
+            vAxis: {
+              title: '自治体',
+              textStyle: {
+                fontSize: 10,
+              },
+            },
+            bar: { groupWidth: '70%' },
+            legend: { position: 'none' },
+          }"
+        />
+      </div>
+    </div>
+
+    <!-- 水道料金ランキング -->
+    <div
+      v-if="activeTab === 'ranking'"
+      class="bg-white p-4 md:p-6 rounded-lg shadow-sm"
+    >
+      <h2 class="text-xl font-semibold mb-4 text-main">
+        全国自治体水道料金ランキング
+      </h2>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">使用量を選択</label>
+        <select
+          v-model="rankingUsage"
+          class="w-full p-2 border rounded"
+          @change="updateRankingData"
+        >
+          <option value="10m3">10m³</option>
+          <option value="15m3">15m³</option>
+          <option value="20m3">20m³</option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">並び順を選択</label>
+        <select v-model="sortOrder" class="w-full p-2 border rounded">
+          <option value="asc">安い順</option>
+          <option value="desc">高い順</option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">自治体名で検索</label>
+        <input
+          v-model="searchCity"
+          type="text"
+          class="w-full p-2 border rounded"
+          placeholder="自治体名を入力してください"
+        />
+        <div class="flex mt-2 space-x-4">
+          <button
+            class="px-4 py-2 bg-main text-white rounded"
+            @click="searchRankingCity"
+          >
+            検索
+          </button>
+          <button
+            v-if="searchedCityResult"
+            class="px-4 py-2 text-white rounded hover:bg-gray-600"
+            style="background-color: #6b7280"
+            @click="clearSearch"
+          >
+            クリア
+          </button>
+        </div>
+      </div>
+      <div v-if="searchedCityResult">
+        <h3 class="text-lg font-semibold mb-2 text-main">検索結果</h3>
+        <div v-if="searchedCityResult.length > 0">
+          <div
+            v-if="searchTargetRank"
+            class="mb-4 p-3 bg-blue-50 rounded border border-blue-200"
+          >
+            <p class="text-center font-medium">
+              「{{ searchCity }}」は全{{ sortedRankingData.length }}自治体中
+              <span class="text-xl font-bold text-main"
+                >{{ searchTargetRank }}位</span
               >
-                <tr bgcolor="#CCCCEE">
-                  <th width="90">使用水量</th>
-                  <th width="90">口径</th>
-                  <th width="90">静岡県</th>
-                  <th width="90">山梨県</th>
-                </tr>
+              です
+            </p>
+          </div>
+          <div class="overflow-x-auto flex justify-center">
+            <table class="min-w-full bg-white border border-gray-200 mb-4">
+              <thead>
                 <tr>
-                  <td width="60" align="right">20㎥</td>
-                  <td width="60" align="right">13mm</td>
-                  <td width="60" align="right" bgcolor="#BDD7EE">2,417円</td>
-                  <td width="60" align="right" bgcolor="#F8CBAD">2,407円</td>
+                  <th
+                    class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+                  >
+                    順位
+                  </th>
+                  <th
+                    class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+                  >
+                    都道府県
+                  </th>
+                  <th
+                    class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+                  >
+                    自治体名
+                  </th>
+                  <th
+                    class="px-4 py-2 border-b border-gray-200 bg-main text-white text-right"
+                  >
+                    水道料金
+                  </th>
                 </tr>
-                <tr>
-                  <td width="60" align="right">20㎥</td>
-                  <td width="60" align="right">20mm</td>
-                  <td width="60" align="right" bgcolor="#F8CBAD">2,651円</td>
-                  <td width="60" align="right" bgcolor="#BDD7EE">2,751円</td>
+              </thead>
+              <tbody>
+                <tr
+                  v-for="(item, idx) in searchedCityResult"
+                  :key="item.city + item.prefecture"
+                  :class="
+                    item.isTarget
+                      ? 'font-bold'
+                      : idx % 2 === 0
+                      ? 'bg-gray-50'
+                      : 'bg-white'
+                  "
+                  :style="item.isTarget ? 'background-color: #FFEDD5;' : ''"
+                >
+                  <td class="px-4 py-2 border-b border-gray-200">
+                    {{ item.rank }}位
+                  </td>
+                  <td class="px-4 py-2 border-b border-gray-200">
+                    {{ item.prefecture }}
+                  </td>
+                  <td class="px-4 py-2 border-b border-gray-200">
+                    {{ item.city }}
+                  </td>
+                  <td class="px-4 py-2 border-b border-gray-200 text-right">
+                    {{ item.fee.toLocaleString() }}円
+                  </td>
                 </tr>
-              </table>
-            </v-row>
-            <v-row class="my-2 mx-2">
-              &emsp;静岡県と山梨県の平均水道料金において使用水量が同じ場合、
-              口径が13mmの時は山梨県の方が静岡県より安くなっています。
-              一方、口径が20mmの時は静岡県の方が山梨県よりも安くなっています。<br /><br />
-              &emsp;次に使用水量の変化による違いを見てみましょう。
-            </v-row>
-            <v-row class="my-2 mx-2">
-              <table
-                border="1"
-                style="border-collapse: collapse"
-                bgcolor="#FFFFFF"
-                cellpadding="100"
-              >
-                <tr bgcolor="#CCCCEE">
-                  <th width="90">使用水量</th>
-                  <th width="90">口径</th>
-                  <th width="90">大阪府</th>
-                  <th width="90">東京都</th>
-                </tr>
-                <tr>
-                  <td width="60" align="right">10㎥</td>
-                  <td width="60" align="right">13mm</td>
-                  <td width="60" align="right" bgcolor="#BDD7EE">1,130円</td>
-                  <td width="60" align="right" bgcolor="#F8CBAD">1,149円</td>
-                </tr>
-                <tr>
-                  <td width="60" align="right">20㎥</td>
-                  <td width="60" align="right">13mm</td>
-                  <td width="60" align="right" bgcolor="#F8CBAD">2,752円</td>
-                  <td width="60" align="right" bgcolor="#BDD7EE">2,630円</td>
-                </tr>
-              </table>
-            </v-row>
-            <v-row class="mt-2 mb-2 mx-4">
-              <div>
-                &emsp;大阪府と東京都の平均水道料金において口径が13mmで同じ場合、
-                使用水量が10㎥の時は大阪府の方が安くなっていますが、
-                使用水量が増えた20㎥の時は東京都の方が安くなっています。<br /><br />
+              </tbody>
+            </table>
+          </div>
+        </div>
+        <div v-else class="text-red-600">
+          該当する自治体が見つかりませんでした。
+        </div>
+      </div>
 
-                &emsp;このように水道料金や口径が異なることによって地域ごとに水道料金が異なることが分かります。<br /><br />
-              </div>
-            </v-row>
-            <v-row class="mt-2 mb-2 mx-4">
-              <div>
-                &emsp;本サイトでは、口径が「13mm」と「20mm」の場合と、水道料金が「10㎥」と「20㎥」の場合の水道料金データを表示しています。
-                (基本料金と超過料金に関しては口径「13mm」のみを使用)
-              </div>
-            </v-row>
-            <v-divider></v-divider>
-            <v-row id="data" class="my-2">
-              <v-col cols="12" md="5">
-                <h1>水道料金データの見方</h1>
-              </v-col>
-            </v-row>
-            <v-row class="my-2 mx-3">
-              <div>
-                &emsp;本サイトで調べられる水道料金データには、次のようなものがあります。<br />
-              </div>
-            </v-row>
-            <v-row class="my-2 mx-3 box2">
-              <ul>
-                <li><strong>日本全体の水道料金データ</strong></li>
-                <ol>
-                  <li>都道府県ごとの平均水道料金ランキング</li>
-                  <li>平均水道料金マップ</li>
-                  <li>事業体(市町村)の水道料金表</li>
-                  <li>全国の事業体の基本料金と超過料金</li>
-                  <li>水道料金比較シミュレーター</li>
-                </ol>
-                <li><strong>各都道府県ごとの水道料金データ</strong></li>
-                <ol>
-                  <li>都道府県内での水道料金ランキング</li>
-                  <li>各事業体の水道料金グラフ</li>
-                  <li>都道府県内の基本料金と超過料金表</li>
-                  <li>水道料金計算シミュレータ</li>
-                </ol>
-              </ul>
-            </v-row>
-            <v-row class="mt-5 mb-3 mx-3">
-              <div>
-                &emsp;本サイトでのデータの見方は以下の画像を参考にしてみてください。<br />
-              </div>
-            </v-row>
-            <v-row justify="center">
-              <v-icon color="#0E2997">mdi-circle</v-icon>
-              <h2><font color="#0E2997">日本全体の水道料金データ</font></h2>
-              <v-col cols="12">
-                <v-img
-                  :src="outlook_japan_pc_src"
-                  class="hidden-sm-and-down"
-                ></v-img>
-                <v-img
-                  :src="outlook_japan_sp_src"
-                  class="hidden-md-and-up"
-                ></v-img>
-              </v-col>
-            </v-row>
-            <v-row class="mx-3 mb-5" justify="center">
-              <div>
-                <v-btn
-                  large
-                  color="#0E2997"
-                  dark
-                  width="100%"
-                  class="btn"
-                  @click="moveData()"
-                >
-                  <v-icon>mdi-chevron-right</v-icon>
-                  <strong>&nbsp;日本全体の水道料金データを見てみよう！</strong>
-                </v-btn>
-              </div>
-            </v-row>
-            <v-row class="mt-10" justify="center">
-              <v-icon color="#0E2997">mdi-circle</v-icon>
-              <h2><font color="#0E2997">都道府県ごとの詳細データ</font></h2>
-              <v-col cols="12">
-                <v-img
-                  :src="outlook_pref_pc"
-                  class="hidden-sm-and-down"
-                ></v-img>
-                <v-img :src="outlook_pref_sp" class="hidden-md-and-up"></v-img>
-              </v-col>
-            </v-row>
-            <v-row class="mx-3 mb-5" justify="center">
-              <div>
-                <v-btn
-                  large
-                  color="#0E2997"
-                  dark
-                  width="100%"
-                  class="btn"
-                  @click="moveDataPref()"
-                >
-                  <v-icon>mdi-chevron-right</v-icon>
-                  <strong>
-                    &nbsp;都道府県ごとの水道料金データを見てみよう！
-                  </strong>
-                </v-btn>
-              </div>
-            </v-row>
-            <v-row class="my-3 mx-3">
-              <div>
-                <br />
-                &emsp;※本サイトで使用する水道料金データは、
-                <a
-                  href="https://www.soumu.go.jp/main_sosiki/c-zaisei/kouei_kessan.html"
-                  target="_blank"
-                >
-                  地方公営企業決算状況調査（通称：決算統計）
-                </a>
-                によって2021年に収集されたデータ を利用しています。<br /><br />
-              </div>
-            </v-row>
-          </v-col>
-        </v-row>
-      </section>
-    </v-container>
+      <!-- 区切り線 -->
+      <hr
+        v-if="searchedCityResult && searchedCityResult.length > 0"
+        class="my-6 border-t border-gray-300"
+      />
+
+      <div class="overflow-x-auto flex justify-center">
+        <table class="min-w-full bg-white border border-gray-200">
+          <thead>
+            <tr>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+              >
+                順位
+              </th>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+              >
+                都道府県
+              </th>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+              >
+                自治体名
+              </th>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-right"
+              >
+                水道料金
+              </th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr
+              v-for="(item, index) in sortedRankingData"
+              :key="index"
+              :class="index % 2 === 0 ? 'bg-gray-50' : 'bg-white'"
+            >
+              <td class="px-4 py-2 border-b border-gray-200">
+                {{ index + 1 }}位
+              </td>
+              <td class="px-4 py-2 border-b border-gray-200">
+                {{ item.prefecture }}
+              </td>
+              <td class="px-4 py-2 border-b border-gray-200">
+                {{ item.city }}
+              </td>
+              <td class="px-4 py-2 border-b border-gray-200 text-right">
+                {{ item.fee.toLocaleString() }}円
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
+
+    <!-- 都道府県平均水道料金ランキング -->
+    <div
+      v-if="activeTab === 'averageRanking'"
+      class="bg-white p-4 md:p-6 rounded-lg shadow-sm"
+    >
+      <h2 class="text-xl font-semibold mb-4 text-main">
+        都道府県平均水道料金ランキング
+      </h2>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">使用量を選択</label>
+        <select
+          v-model="averageUsage"
+          class="w-full p-2 border rounded"
+          @change="updateAverageRankingData"
+        >
+          <option value="10m3">10m³</option>
+          <option value="15m3">15m³</option>
+          <option value="20m3">20m³</option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">並び順を選択</label>
+        <select
+          v-model="averageSortOrder"
+          class="w-full p-2 border rounded"
+          @change="updateAverageRankingData"
+        >
+          <option value="asc">安い順</option>
+          <option value="desc">高い順</option>
+        </select>
+      </div>
+      <div class="mb-4">
+        <label class="block mb-2 font-medium">都道府県名で検索</label>
+        <input
+          v-model="averageGraphSearchPrefecture"
+          type="text"
+          class="w-full p-2 border rounded"
+          placeholder="都道府県名を入力してください"
+        />
+        <div class="flex mt-2 space-x-4">
+          <button
+            class="px-6 py-2 bg-main text-white rounded hover:bg-main-dark"
+            @click="searchAverageGraphPrefecture"
+          >
+            検索
+          </button>
+          <button
+            v-if="averageGraphHighlightedPrefecture"
+            class="px-6 py-2 text-white rounded hover:bg-gray-600"
+            style="background-color: #6b7280"
+            @click="clearAverageGraphSearch"
+          >
+            クリア
+          </button>
+        </div>
+      </div>
+      <div
+        v-if="averageGraphHighlightedPrefecture"
+        class="mb-3 p-3 bg-blue-50 rounded-lg border border-blue-200"
+      >
+        <p class="text-center font-medium">
+          「{{ averageGraphHighlightedPrefecture }}」の平均水道料金:
+          <span class="text-xl font-bold text-main">
+            {{ averageGraphHighlightedFee.toLocaleString() }}円
+          </span>
+          ({{ averageGraphHighlightedRank }}/47位)
+        </p>
+      </div>
+      <div
+        v-if="averageRankingData.length"
+        style="min-height: 500px; margin-bottom: 2rem"
+      >
+        <GChart
+          type="ColumnChart"
+          :data="averageChartData"
+          :options="{
+            title: '都道府県平均水道料金',
+            chartArea: { width: '70%', height: '70%' },
+            height: 600,
+            hAxis: {
+              title: '都道府県',
+              slantedText: true,
+              slantedTextAngle: 45,
+              textStyle: {
+                fontSize: 12,
+              },
+              showTextEvery: 1,
+            },
+            vAxis: {
+              title: '料金 (円)',
+              minValue: 0,
+            },
+            bar: { groupWidth: '60%' },
+            legend: { position: 'none' },
+          }"
+        />
+      </div>
+      <hr class="my-4 border-t border-gray-300" />
+      <div class="overflow-x-auto flex justify-center">
+        <table class="min-w-full bg-white border border-gray-200">
+          <thead>
+            <tr>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+              >
+                順位
+              </th>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-left"
+              >
+                都道府県
+              </th>
+              <th
+                class="px-4 py-2 border-b border-gray-200 bg-main text-white text-right"
+              >
+                平均水道料金
+              </th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr
+              v-for="(item, index) in sortedAverageRankingData"
+              :key="index"
+              :class="index % 2 === 0 ? 'bg-gray-50' : 'bg-white'"
+            >
+              <td class="px-4 py-2 border-b border-gray-200">
+                {{ index + 1 }}位
+              </td>
+              <td class="px-4 py-2 border-b border-gray-200">
+                {{ item.prefecture }}
+              </td>
+              <td class="px-4 py-2 border-b border-gray-200 text-right">
+                {{ Number(item.averageFee).toFixed(1).toLocaleString() }}円
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+      <div class="text-right mb-4">
+        <button
+          class="px-4 py-2 bg-main text-white rounded hover:bg-main-dark"
+          @click="downloadCSV"
+        >
+          CSVダウンロード
+        </button>
+      </div>
+    </div>
+    <hr class="my-8 border-t border-gray-300" />
+    <!-- 水道料金説明ページへのリンク -->
+    <div class="mt-12 mb-8 py-4 text-center">
+      <a
+        href="/charge"
+        class="px-8 py-3 bg-main text-white rounded-lg shadow-md hover:bg-main-dark transition-colors font-medium"
+        style="text-decoration: none"
+      >
+        水道料金の計算方法について詳しく知る →
+      </a>
+      <p class="mt-6 mb-8 text-sm text-gray-600">
+        水道料金の仕組みについて解説しています。<br />
+        水道料金がどのように計算されているのか知りたい方は上のボタンを押してみてください。
+      </p>
+    </div>
   </div>
 </template>
 
+
 <script>
-import BreadCrumb from '~/components/atoms/BreadCrumb.vue'
+import Vue from 'vue'
+import VueGoogleCharts from 'vue-google-charts'
+
+Vue.use(VueGoogleCharts)
 
 export default {
-  name: 'ChargeHome',
-  components: {
-    BreadCrumb,
-  },
+  name: 'ChargeFix',
+
   data() {
     return {
-      full_cost: require('@/assets/img/charge_intro/full_cost.png'),
-      charge_system: require('@/assets/img/charge_intro/charge_system.png'),
-      change_graph: require('@/assets/img/charge_intro/ratio_change_graph.png'),
-      ratio_system: require('@/assets/img/charge_intro/ratio_system.png'),
-      outlook_japan_pc_src: require('@/assets/img/outlook_charge/japan_pc.png'),
-      outlook_japan_sp_src: require('@/assets/img/outlook_charge/japan_sp.png'),
-      outlook_pref_pc: require('@/assets/img/outlook_charge/pref_pc.png'),
-      outlook_pref_sp: require('@/assets/img/outlook_charge/pref_sp.png'),
+      waterRatesData: [],
+      selectedPrefecture: '東京都',
+      comparePrefecture: '神奈川県',
+      selectedCity: '東京都',
+      compareCity: '横浜市',
+      waterUsage: 20,
+      rankingUsage: '20m3',
+      sortOrder: 'asc',
+      activeTab: 'compare',
+      rankingData: [],
+      prefectureOrder: [
+        '北海道',
+        '青森県',
+        '岩手県',
+        '宮城県',
+        '秋田県',
+        '山形県',
+        '福島県',
+        '茨城県',
+        '栃木県',
+        '群馬県',
+        '埼玉県',
+        '千葉県',
+        '東京都',
+        '神奈川県',
+        '新潟県',
+        '富山県',
+        '石川県',
+        '福井県',
+        '山梨県',
+        '長野県',
+        '岐阜県',
+        '静岡県',
+        '愛知県',
+        '三重県',
+        '滋賀県',
+        '京都府',
+        '大阪府',
+        '兵庫県',
+        '奈良県',
+        '和歌山県',
+        '鳥取県',
+        '島根県',
+        '岡山県',
+        '広島県',
+        '山口県',
+        '徳島県',
+        '香川県',
+        '愛媛県',
+        '高知県',
+        '福岡県',
+        '佐賀県',
+        '長崎県',
+        '熊本県',
+        '大分県',
+        '宮崎県',
+        '鹿児島県',
+        '沖縄県',
+      ],
+      selectedGraphPrefecture: '東京都',
+      selectedUsageFee: '20m3',
+      searchCity: '',
+      searchedCityResult: null,
+      searchTargetRank: null,
+      isMobile: false,
+      graphSearchCity: '',
+      graphHighlightedCity: null,
+      chartData: [],
+      graphHighlightedCityFee: null,
+      averageUsage: '20m3',
+      averageRankingData: [],
+      averageChartData: [],
+      averageSortOrder: 'asc',
+      averageGraphSearchPrefecture: '',
+      averageGraphHighlightedPrefecture: null,
+      averageGraphHighlightedFee: null,
+      averageGraphHighlightedRank: null,
     }
   },
+
   computed: {
-    breadcrumbs() {
-      return [
-        {
-          name: 'ホーム',
-          path: '/',
-        },
-        {
-          name: '水道料金',
-        },
-      ]
+    availablePrefectures() {
+      const availablePrefSet = new Set(
+        this.waterRatesData.map((data) => data.prefecture)
+      )
+      return this.prefectureOrder.filter((pref) => availablePrefSet.has(pref))
+    },
+
+    availableCitiesInPrefecture() {
+      return this.waterRatesData
+        .filter((data) => data.prefecture === this.selectedPrefecture)
+        .map((data) => data.municipality)
+        .sort()
+    },
+
+    availableCitiesInComparePrefecture() {
+      return this.waterRatesData
+        .filter((data) => data.prefecture === this.comparePrefecture)
+        .map((data) => data.municipality)
+        .sort()
+    },
+
+    calculateNationalAverage() {
+      if (!this.waterRatesData.length) return 0
+      const total = this.waterRatesData.reduce(
+        (sum, data) => sum + this.calculateWaterBill(data),
+        0
+      )
+      return Math.round(total / this.waterRatesData.length)
+    },
+
+    bill() {
+      return this.calculateWaterBill(this.selectedCityData)
+    },
+
+    compareBill() {
+      return this.calculateWaterBill(this.compareCityData)
+    },
+
+    selectedCityData() {
+      return this.waterRatesData.find(
+        (data) => data.municipality === this.selectedCity
+      )
+    },
+
+    compareCityData() {
+      return this.waterRatesData.find(
+        (data) => data.municipality === this.compareCity
+      )
+    },
+
+    sortedRankingData() {
+      if (this.sortOrder === 'asc') {
+        return [...this.rankingData].sort((a, b) => a.fee - b.fee)
+      } else {
+        return [...this.rankingData].sort((a, b) => b.fee - a.fee)
+      }
+    },
+
+    sortedAverageRankingData() {
+      if (this.averageSortOrder === 'asc') {
+        return [...this.averageRankingData].sort(
+          (a, b) => a.averageFee - b.averageFee
+        )
+      } else {
+        return [...this.averageRankingData].sort(
+          (a, b) => b.averageFee - a.averageFee
+        )
+      }
     },
   },
-  methods: {
-    moveData() {
-      this.$router.push('/charge/data/')
+
+  watch: {
+    selectedPrefecture(newPref) {
+      if (!this.availablePrefectures.includes(newPref)) {
+        this.selectedCity = ''
+        return
+      }
+      if (!this.availableCitiesInPrefecture.includes(this.selectedCity)) {
+        this.selectedCity = this.availableCitiesInPrefecture[0] || ''
+      }
     },
-    moveDataPref() {
-      this.$router.push('/charge/data/select_pref')
+    comparePrefecture(newPref) {
+      if (!this.availablePrefectures.includes(newPref)) {
+        this.compareCity = ''
+        return
+      }
+      if (!this.availableCitiesInComparePrefecture.includes(this.compareCity)) {
+        this.compareCity = this.availableCitiesInComparePrefecture[0] || ''
+      }
+    },
+    selectedGraphPrefecture() {
+      // 都道府県が変わったら検索結果をリセット
+      this.graphHighlightedCity = null
+      this.graphHighlightedCityFee = null
+      this.graphSearchCity = ''
+      this.updateGraph()
+    },
+    selectedUsageFee() {
+      // 使用量が変わったら、ハイライト自治体の料金も更新
+      if (this.graphHighlightedCity) {
+        const foundCity = this.waterRatesData.find(
+          (data) =>
+            data.prefecture === this.selectedGraphPrefecture &&
+            data.municipality === this.graphHighlightedCity
+        )
+        if (foundCity) {
+          this.graphHighlightedCityFee =
+            foundCity.usage_fee[this.selectedUsageFee]
+        }
+      }
+      this.updateGraph()
+    },
+    rankingUsage: 'updateRankingData',
+    sortOrder() {
+      // ソート順が変更されたときは計算プロパティが自動的に再計算される
+    },
+    activeTab(newTab) {
+      if (newTab === 'ranking') {
+        this.updateRankingData()
+      } else if (newTab === 'graph') {
+        this.updateGraph()
+      } else if (newTab === 'averageRanking') {
+        this.updateAverageRankingData()
+      }
+    },
+  },
+
+  async created() {
+    try {
+      const module = await import(
+        '@/assets/mockup_data/ChargeData/water_rates.json'
+      )
+      this.waterRatesData = module.default
+
+      if (
+        !this.waterRatesData.some(
+          (data) =>
+            data.prefecture === '東京都' && data.municipality === '東京都'
+        )
+      ) {
+        this.selectedCity = this.availableCitiesInPrefecture[0] || ''
+      }
+      if (
+        !this.waterRatesData.some(
+          (data) =>
+            data.prefecture === '神奈川県' && data.municipality === '横浜市'
+        )
+      ) {
+        this.compareCity = this.availableCitiesInComparePrefecture[0] || ''
+      }
+
+      this.updateRankingData()
+      this.updateGraph()
+      this.updateAverageRankingData()
+      this.checkMobile()
+    } catch (error) {
+      this.$emit('error', 'データの読み込みに失敗しました')
+    }
+  },
+
+  mounted() {
+    window.addEventListener('resize', this.checkMobile)
+    this.checkMobile()
+  },
+
+  beforeDestroy() {
+    window.removeEventListener('resize', this.checkMobile)
+  },
+
+  methods: {
+    checkMobile() {
+      this.isMobile = window.innerWidth < 768
+    },
+
+    calculateWaterBill(cityData) {
+      if (!cityData) return 0
+
+      const {
+        basic_rate: basicRate,
+        excess_rate: excessRate,
+        basic_volume: basicVolume,
+        meter_usage: meterUsage,
+      } = cityData
+      const usage = this.waterUsage
+
+      if (usage <= basicVolume) {
+        return basicRate + meterUsage
+      }
+
+      return basicRate + (usage - basicVolume) * excessRate + meterUsage
+    },
+
+    updateRankingData() {
+      this.rankingData = this.waterRatesData.map((data) => {
+        return {
+          prefecture: data.prefecture,
+          city: data.municipality,
+          fee: data.usage_fee[this.rankingUsage] || 0,
+        }
+      })
+    },
+
+    updateGraph() {
+      const selectedData = this.waterRatesData.filter(
+        (data) => data.prefecture === this.selectedGraphPrefecture
+      )
+
+      // データの並び替え（料金の降順）
+      selectedData.sort((a, b) => {
+        return (
+          b.usage_fee[this.selectedUsageFee] -
+          a.usage_fee[this.selectedUsageFee]
+        )
+      })
+
+      // Google Chartsのためのデータ形式に変換
+      const dataTable = [['自治体', '料金', { role: 'style' }]]
+
+      // データを追加（ハイライト対象はオレンジ色に）
+      selectedData.forEach((data) => {
+        const color =
+          data.municipality === this.graphHighlightedCity
+            ? '#FF9800' // オレンジ
+            : '#0E2997' // 青
+
+        dataTable.push([
+          data.municipality,
+          data.usage_fee[this.selectedUsageFee],
+          color,
+        ])
+      })
+
+      this.chartData = dataTable
+    },
+
+    searchRankingCity() {
+      const keyword = this.searchCity.trim()
+      if (!keyword) {
+        this.searchedCityResult = []
+        this.searchTargetRank = null
+        return
+      }
+      // ランキングデータを取得
+      const data = this.sortedRankingData
+      // 検索
+      const idx = data.findIndex((item) => item.city.includes(keyword))
+      if (idx === -1) {
+        this.searchedCityResult = []
+        this.searchTargetRank = null
+        return
+      }
+
+      // 検索対象の順位を保存
+      this.searchTargetRank = idx + 1
+
+      // 前後2件を取得（計最大5件）
+      const start = Math.max(0, idx - 2)
+      const end = Math.min(data.length, idx + 3)
+      this.searchedCityResult = data.slice(start, end).map((item, i) => ({
+        ...item,
+        rank: start + i + 1,
+        isTarget: start + i === idx,
+      }))
+    },
+
+    clearSearch() {
+      this.searchCity = ''
+      this.searchedCityResult = null
+      this.searchTargetRank = null
+    },
+
+    searchGraphCity() {
+      const keyword = this.graphSearchCity.trim()
+      if (!keyword) {
+        this.graphHighlightedCity = null
+        this.updateGraph()
+        return
+      }
+
+      // 現在選択中の都道府県の自治体を検索
+      const selectedData = this.waterRatesData.filter(
+        (data) => data.prefecture === this.selectedGraphPrefecture
+      )
+
+      // 検索
+      const foundCity = selectedData.find((data) =>
+        data.municipality.includes(keyword)
+      )
+
+      if (!foundCity) {
+        alert('該当する自治体が見つかりませんでした')
+        this.graphHighlightedCity = null
+      } else {
+        // 検索対象の自治体を保存
+        this.graphHighlightedCity = foundCity.municipality
+        this.graphHighlightedCityFee =
+          foundCity.usage_fee[this.selectedUsageFee]
+      }
+
+      // グラフを更新
+      this.updateGraph()
+    },
+
+    clearGraphSearch() {
+      this.graphSearchCity = ''
+      this.graphHighlightedCity = null
+      this.graphHighlightedCityFee = null
+      this.updateGraph()
+    },
+
+    updateAverageRankingData() {
+      const averageData = this.waterRatesData.reduce((acc, data) => {
+        if (!acc[data.prefecture]) {
+          acc[data.prefecture] = { total: 0, count: 0 }
+        }
+        acc[data.prefecture].total += data.usage_fee[this.averageUsage]
+        acc[data.prefecture].count += 1
+        return acc
+      }, {})
+
+      this.averageRankingData = Object.entries(averageData).map(
+        ([prefecture, { total, count }]) => ({
+          prefecture,
+          averageFee: total / count,
+        })
+      )
+
+      if (this.averageSortOrder === 'asc') {
+        this.averageRankingData.sort((a, b) => a.averageFee - b.averageFee)
+      } else {
+        this.averageRankingData.sort((a, b) => b.averageFee - a.averageFee)
+      }
+
+      this.averageChartData = [['都道府県', '平均料金', { role: 'style' }]]
+      this.averageRankingData.forEach((data) => {
+        this.averageChartData.push([
+          data.prefecture,
+          parseFloat(Number(data.averageFee).toFixed(1)),
+          '#0E2997',
+        ])
+      })
+    },
+
+    searchAverageGraphPrefecture() {
+      const keyword = this.averageGraphSearchPrefecture.trim()
+      if (!keyword) {
+        this.averageGraphHighlightedPrefecture = null
+        this.averageGraphHighlightedFee = null
+        this.averageGraphHighlightedRank = null
+        this.updateAverageGraph()
+        return
+      }
+
+      const foundIndex = this.sortedAverageRankingData.findIndex((data) =>
+        data.prefecture.includes(keyword)
+      )
+
+      if (foundIndex === -1) {
+        alert('該当する都道府県が見つかりませんでした')
+        this.averageGraphHighlightedPrefecture = null
+        this.averageGraphHighlightedFee = null
+        this.averageGraphHighlightedRank = null
+      } else {
+        const foundPrefecture = this.sortedAverageRankingData[foundIndex]
+        this.averageGraphHighlightedPrefecture = foundPrefecture.prefecture
+        this.averageGraphHighlightedFee = foundPrefecture.averageFee
+        this.averageGraphHighlightedRank = foundIndex + 1
+      }
+
+      this.updateAverageGraph()
+    },
+
+    clearAverageGraphSearch() {
+      this.averageGraphSearchPrefecture = ''
+      this.averageGraphHighlightedPrefecture = null
+      this.averageGraphHighlightedFee = null
+      this.averageGraphHighlightedRank = null
+      this.updateAverageGraph()
+    },
+
+    updateAverageGraph() {
+      this.averageChartData = [['都道府県', '平均料金', { role: 'style' }]]
+      this.averageRankingData.forEach((data) => {
+        const color =
+          data.prefecture === this.averageGraphHighlightedPrefecture
+            ? '#FF9800' // オレンジ
+            : '#0E2997' // 青
+
+        this.averageChartData.push([
+          data.prefecture,
+          parseFloat(Number(data.averageFee).toFixed(1)),
+          color,
+        ])
+      })
+    },
+
+    downloadCSV() {
+      const headers = ['順位', '都道府県', '平均水道料金']
+      const rows = this.sortedAverageRankingData.map((item, index) => [
+        `${index + 1}位`,
+        item.prefecture,
+        `${item.averageFee.toFixed(1)}円`,
+      ])
+
+      const csvContent = [headers, ...rows].map((e) => e.join(',')).join('\n')
+
+      const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' })
+      const link = document.createElement('a')
+      link.href = URL.createObjectURL(blob)
+      link.setAttribute('download', 'average_water_fees.csv')
+      document.body.appendChild(link)
+      link.click()
+      document.body.removeChild(link)
     },
   },
 }
 </script>
 
 <style scoped>
-.c-p {
-  cursor: pointer;
+.p-4 {
+  padding: 1rem;
 }
 
-.charge-intro {
-  background: rgba(242 242 242);
-  width: 100%;
+.max-w-6xl {
+  max-width: 72rem;
 }
 
-h1 {
-  position: relative;
-  background: #f1f8ff;
-  padding: 0.25em 0.5em;
-  border-left: solid 2em #5c9ee7;
-  width: 100%;
+.mx-auto {
+  margin-left: auto;
+  margin-right: auto;
 }
 
-h1::before {
-  content: url('@/assets/img/icon/water-check.png');
-  position: absolute;
-  color: white;
-  left: -1.8em;
-  top: 60%;
-  transform: translateY(-50%);
+.bg-gray-50 {
+  background-color: #f9fafb;
 }
 
-.box2 ul {
+.rounded-lg {
+  border-radius: 0.5rem;
+}
+
+.shadow {
+  box-shadow: 0 1px 3px rgb(0 0 0 / 10%), 0 1px 2px rgb(0 0 0 / 6%);
+}
+
+.text-2xl {
+  font-size: 1.5rem;
+  line-height: 2rem;
+}
+
+.font-bold {
+  font-weight: 700;
+}
+
+.text-center {
+  text-align: center;
+}
+
+.mb-6 {
+  margin-bottom: 1.5rem;
+}
+
+.text-main {
   color: #0e2997;
-  border: dashed 2px #668ad8;
-  background: #f1f8ff;
-  padding: 0.5em 0.5em 0.5em 2em;
-  width: 110%;
 }
 
-.box2 ul li {
-  line-height: 1.5;
-  padding: 0.5em 0;
+.text-main-dark {
+  color: #081d6e;
 }
 
-.box1 {
-  padding: 0.5em 1em;
-  margin: 0.5em 0;
-  font-weight: bold;
-  border: solid 3px #000;
+.bg-main {
+  background-color: #0e2997;
 }
 
-.box1 p {
-  margin: 0;
-  padding: 0;
+.bg-main-light {
+  background-color: #4e67c9;
 }
 
-td {
-  padding: 5px;
+.bg-light-blue {
+  background-color: #e1f5fe;
 }
 
-th {
-  padding: 5px;
+.bg-light-orange {
+  background-color: #fff3e0;
 }
 
-.table-wrap {
-  overflow-x: scroll;
+.border-main-light {
+  border-color: #4e67c9;
 }
 
-.btn {
-  font-size: 16px;
+/* タブメニューのスタイル */
+.flex {
+  display: flex;
 }
 
-@media screen and (max-width: 600px) {
-  .btn {
-    font-size: 12px;
+.flex-wrap {
+  flex-wrap: wrap;
+}
+
+.bg-white {
+  background-color: #fff;
+}
+
+.shadow-sm {
+  box-shadow: 0 1px 2px rgb(0 0 0 / 5%);
+}
+
+button {
+  border: none;
+  cursor: pointer;
+  transition: all 0.3s ease;
+}
+
+button.px-4 {
+  padding-left: 1rem;
+  padding-right: 1rem;
+}
+
+button.py-2 {
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+}
+
+.font-medium {
+  font-weight: 500;
+}
+
+.text-white {
+  color: #fff;
+}
+
+.bg-gray-200 {
+  background-color: #e5e7eb;
+}
+
+/* グリッドレイアウト */
+.grid {
+  display: grid;
+}
+
+.grid-cols-1 {
+  grid-template-columns: repeat(1, minmax(0, 1fr));
+}
+
+/* メディアクエリの修正 */
+@media (min-width: 768px) {
+  .md-grid-cols-2 {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
   }
+}
+
+.gap-6 {
+  gap: 1.5rem;
+}
+
+/* フォームコントロール */
+.block {
+  display: block;
+}
+
+.w-full {
+  width: 100%;
+}
+
+input[type='range'] {
+  width: 100%;
+  margin: 0.5rem 0;
+}
+
+select,
+input {
+  padding: 0.5rem;
+  border: 1px solid #d1d5db;
+  border-radius: 0.375rem;
+  background-color: #fff;
+}
+
+/* 計算結果エリア */
+.text-lg {
+  font-size: 1.125rem;
+  line-height: 1.75rem;
+}
+
+.text-3xl {
+  font-size: 1.875rem;
+  line-height: 2.25rem;
+}
+
+.text-sm {
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+}
+
+.text-gray-600 {
+  color: #4b5563;
+}
+
+.mt-2 {
+  margin-top: 0.5rem;
+}
+
+/* ボタングループ */
+.space-x-4 > * + * {
+  margin-left: 1rem;
+}
+
+.hover-bg-main:hover {
+  background-color: #081d6e;
+}
+
+.hover-bg-green-700:hover {
+  background-color: #15803d;
+}
+
+.bg-green-600 {
+  background-color: #16a34a;
+}
+
+/* レスポンシブデザインの調整 */
+@media (max-width: 640px) {
+  .p-4 {
+    padding: 0.75rem;
+  }
+
+  .text-2xl {
+    font-size: 1.25rem;
+  }
+
+  .radio-tab {
+    padding: 0.75rem 0.5rem;
+    font-size: 0.875rem;
+  }
+}
+
+/* アニメーション */
+.transition {
+  transition-property: background-color, border-color, color, fill, stroke,
+    opacity, box-shadow, transform;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 150ms;
+}
+
+/* タブの選択状態をより明確に */
+.radio-tab[class*='bg-main']::after {
+  content: '';
+  position: absolute;
+  bottom: -2px;
+  left: 0;
+  width: 100%;
+  height: 2px;
+  background-color: #0e2997;
+}
+
+/* 計算結果の強調表示 */
+.text-3xl.font-bold {
+  text-shadow: 1px 1px 2px rgb(0 0 0 / 10%);
+}
+
+.comparison-container {
+  display: grid;
+  gap: 1.5rem;
+  grid-template-columns: 1fr;
+}
+
+@media (min-width: 768px) {
+  .comparison-container {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+/* グラフの外側のスペースをさらに大きくする */
+.bg-white.p-6.rounded-lg.shadow-sm {
+  padding: 3rem;
+  margin: 1rem 0 3rem;
+}
+
+/* 見出しとプルダウンの余白を小さくする */
+.mb-4 {
+  margin-bottom: 1rem;
+}
+
+.block.mb-2 {
+  margin-bottom: 0.5rem;
+}
+
+select {
+  appearance: none;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='%230E2997'%3E%3Cpath d='M7 10l5 5 5-5z'/%3E%3C/svg%3E");
+  background-repeat: no-repeat;
+  background-position: right 8px center;
+  background-size: 16px;
+  padding-right: 30px;
 }
 </style>
